refactor(vue): consolidate provider to styled-system directory
- Move provider to styled-system/provider.ts to match React structure
- Simplify provider to just provide SystemContext (Panda CSS handles styling)
- Remove old chakra-provider.ts, use-chakra.ts, use-style-config.ts
- Remove test files and vitest config (aligned with React package)
- Update exports in index.ts and styled-system/index.ts
- Clean up package.json (remove test script and test devDependencies)
