+++ packages/vue/package.json
@@ -36,8 +36,7 @@
     "build": "tsx ../../scripts/build/main.ts --dts",
     "dev": "pnpm build:fast --watch",
     "clean": "rimraf dist .turbo",
-    "typecheck": "tsc --noEmit",
-    "test": "vitest"
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
     "@chakra-ui/system-core": "workspace:*"
@@ -46,9 +45,6 @@
     "vue": "^3.3.0"
   },
   "devDependencies": {
-    "@vue/test-utils": "^2.4.6",
-    "jsdom": "^26.1.0",
-    "vitest": "^3.2.4",
     "vue": "^3.5.13"
   },
   "exports": {
